Fixed linking issue.
