+++ src/base_classes/active_space_method.cc
@@ -35,10 +35,7 @@
 #include "sci/aci.h"
 #include "sci/asci.h"
 #include "sci/fci_mo.h"
-#include "pci/pci.h"
-#include "pci/pci_hashvec.h"
 #include "pci/ewci.h"
-#include "pci/pci_simple.h"
 #include "sci/aci_sci.h"
 #include "ci_ex_states/excited_state_solver.h"
 
@@ -89,18 +86,18 @@ std::unique_ptr<ActiveSpaceMethod> make_active_space_method(
         solver = std::make_unique<ASCI>(state, nroot, scf_info, options, mo_space_info, as_ints);
     } else if (type == "CASSCF") {
         solver = std::make_unique<CASSCF>(state, nroot, scf_info, options, mo_space_info, as_ints);
-    } else if (type == "PCI") {
-        // TODO modify pci code to compute multiple roots under new framework
-        solver =
-            std::make_unique<ProjectorCI>(state, nroot, scf_info, options, mo_space_info, as_ints);
-    } else if (type == "PCI_SIMPLE") {
-        // TODO modify pci code to compute multiple roots under new framework
-        solver = std::make_unique<ProjectorCI_Simple>(state, nroot, scf_info, options,
-                                                      mo_space_info, as_ints);
-    } else if (type == "PCI_HASHVEC") {
-        // TODO modify pci code to compute multiple roots under new framework
-        solver = std::make_unique<ProjectorCI_HashVec>(state, nroot, scf_info, options,
-                                                       mo_space_info, as_ints);
+//    } else if (type == "PCI") {
+//        // TODO modify pci code to compute multiple roots under new framework
+//        solver =
+//            std::make_unique<ProjectorCI>(state, nroot, scf_info, options, mo_space_info, as_ints);
+//    } else if (type == "PCI_SIMPLE") {
+//        // TODO modify pci code to compute multiple roots under new framework
+//        solver = std::make_unique<ProjectorCI_Simple>(state, nroot, scf_info, options,
+//                                                      mo_space_info, as_ints);
+//    } else if (type == "PCI_HASHVEC") {
+//        // TODO modify pci code to compute multiple roots under new framework
+//        solver = std::make_unique<ProjectorCI_HashVec>(state, nroot, scf_info, options,
+//                                                       mo_space_info, as_ints);
     } else if (type == "EWCI") {
         // TODO modify pci code to compute multiple roots under new framework
 //        solver = std::make_unique<ElementwiseCI>(state, nroot, scf_info, options, mo_space_info,
